backend/requirements.txt
# Recensio - Backend Dependencies
# Compatible versions for Pydantic AI

# Core Framework - updated for compatibility
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
pydantic>=2.10.0
python-dotenv>=1.0.0


# Cache/Redis
redis>=5.0.0
hiredis>=2.2.0

# HTTP Client - compatible with all AI dependencies
httpx>=0.27.0

# AI and content processing with Pydantic AI
pydantic-ai==1.1.0
anthropic>=0.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Logging
loguru>=0.7.2

# Modern configuration and validation
pydantic[email]>=2.10.0

# Development
pytest==7.4.3